Add EC2 remote setup script for an EC2 API server
This includes a (slightly rough) setup script for using EC2 instances. It mimics the existing setup_aws.py script in finding/creating an RDS instance but it replaces the code that spins up a Lambda service. Instead it creates an RDS config file and copies it and all other requisite files to the EC2 server, ssh's in, and restarts the service. The API endpoint is then the public address of the EC2 instance.